Update main.py
fixed BUG: When someone quits from IRC the bot tries to get the command from t[2]. This didn't exist when the bot received ["nick", "QUIT"]. This has been fixed now. If there are any more bugs pls commit to the main repo or if you want to add a feature why not fork?